Erste Warenproduktion: Gebäude produzieren nun Waren und erhöhen damit den internen Lagerbestand im Gebäude, bis das Lager voll is - UpdateableObject als weitere Basisklasse für Strukturen hinzugefügt - Lagerbestände sind nun intern double-Werte, der Benutzer soll aber nur Ganzzahlen sehen

+++ src/cpp/gui/GuiUtils.h
@@ -18,7 +18,7 @@ public:
      * @param goodsType welches Gut
      * @param inventory Lagerbestand, der ausgegeben werden soll (oder -1, wenn keine Anzeige erwünscht)
      */
-    static void drawGoodsBox(int x, int y, GoodsType goodsType, int inventory);
+    static void drawGoodsBox(int x, int y, GoodsType goodsType, double inventory);
 
 };
